refactor: all project

+++ src/main/java/ru/hackathon/alphabank/recommendsystem/model/ClientData.java
@@ -5,8 +5,8 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 import java.util.List;
 
 public record ClientData(
-        @JsonProperty("clientId") int clientId,
-        @JsonProperty("organizationId") int organizationId,
+        @JsonProperty("clientId") long clientId,
+        @JsonProperty("organizationId") long organizationId,
         @JsonProperty("segment") String segment, // "Малый бизнес", "Средний бизнес", "Крупный бизнес"
         @JsonProperty("role") String role, // "ЕИО", "Сотрудник"
         @JsonProperty("organizations") int organizations,
